+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- nOS-V model version increased to 2.3.0.
+- Prevent accidental use of nOS-V traces without required events for the
+  breakdown model using the `nosv.can_breakdown` attribute.
+
+### Added
+
+- Add support for nOS-V progressing events VP{pra}.
+- Add breakdown model for nOS-V.
+- New API to manage stream metadata `ovni_attr_*()`.
+
 ## [1.9.1] - 2024-05-10
 
 ### Fixed